CHROMIUM: mm/mglru: Improve isolated page access bit harvesting
During eviction, between the time when a particular folio is isolated
and when it is actually reclaimed, the folio is not in any generation.
Since page access bit harvesting can be done concurrently with eviction,
folio_update_gen() needs special handling for these folios. This is
currently done by clearing the folio's referenced bit when it is
isolated, and having folio_update_gen() mark any isolated folios when
they are accessed.

Instead, we can take advantage of the fact that not all values of page
flag's generation bits are needed to store the normal generations. The
three previously unused values are used to mark a folio as isolated and
count between 0 and 2 harvested accessed bits. This allows the access
count to be precisely communicated to folio_check_references().

BUG=b:342531825
UPSTREAM-TASK=b:357764239
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: Ifc18002e90942911e473019c06cd8bafff742454
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5922298
Reviewed-by: Suleiman Souhlal <suleiman@google.com>