CHROMIUM: mm/mglru: slow down promotion through page tables
Add the option for MGLRU to require two accesses via page tables before
promoting a folio to the newest generation. The folio's referenced bit
is used to record whether or not a page has already been accessed. When
harvesting accessed bits, if that bit isn't set, set it and only promote
the page by one generation. Otherwise promote the page to the newest
generation.

This can be helpful when the working set is large relative to the amount
of system memory. Pages that are only used infrequently will likely be
evicted before they are used again. Slowing down promotion
de-prioritizes such pages and makes it less likely that pages actually
in the working set will be evicted. This behavior resembles the old
LRU's inactive -> active promotion process.

The new behavior is guarded by the new LRU_GEN_AGGRESSIVE_MM bit in the
MGLRU's enable bitmask in sysfs. The default enabled value of that flag
preserves the historic behavior of MGLRU, and disabling the flag enables
the new, more conservative behavior.

Note that accesses through file descriptors will also set the referenced
bit. This is fine, since it's still two accesses, so the page should be
promoted to the most recent generation.

BUG=b:342531825
UPSTREAM-TASK=b:357764239
TEST=tast run DUT multivm.Lifecycle.host_with_bg_arc_vmmms

Change-Id: Id844dee33e76261d7dc2defa450e138da4e06b1f
Signed-off-by: David Stevens <stevensd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5922299
Reviewed-by: Suleiman Souhlal <suleiman@google.com>