CHROMIUM: iwl7000: mvm: clean up ROC on failure
If the firmware fails to start the session protection, then we
do call iwl_mvm_roc_finished() here, but that won't do anything
at all because IWL_MVM_STATUS_ROC_P2P_RUNNING was never set.
Set IWL_MVM_STATUS_ROC_P2P_RUNNING in the failure/stop path.
If it started successfully before, it's already set, so that
doesn't matter, and if it didn't start it needs to be set to
clean up.

Not doing so will lead to a WARN_ON() later on a fresh remain-
on-channel, since the link is already active when activated as
it was never deactivated.

Change-Id: I67f5dac742170dd937f11e4d4f937f45f71b7cb4
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 8d781e1e73eb77f340d3c4be580fc11709feaf67