CHROMIUM: arm64: dts: mt8186: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading
the power, the display power is calculated in watts due to
unsynchronized reading and writing. Use virtual battery 0xf
to communicate with EC can solve this.

BUG=b:372336821
TEST=check battery charging and discharge function
UPSTREAM-TASK=b:335337385

Change-Id: Ia5c36ba848b379433d4b6f1fa6223aba96c5d6eb
Signed-off-by: Geoffrey Chien <geoffrey_chien@pegatron.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5952745
Reviewed-by: Herbert Wu <herbert1_wu@pegatron.corp-partner.google.com>
Commit-Queue: Shawn Ku <shawnku@chromium.org>
Tested-by: Herbert Wu <herbert1_wu@pegatron.corp-partner.google.com>
Reviewed-by: Shawn Ku <shawnku@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>