CHROMIUM: ASoC: SOF: Check runtime for nullity in rx IPCs
There is a failure case where the DSP reports the status of a stream
after the kernel has given up and closed out the pcm. When this happens
the runtime parameter on the substream is nulled out by ASoC in the
pcm_close call back. This results in a NULL pointer derefence if not
checked.

BUG=b:374669573
TEST=Build test
UPSTREAM-TASK=b:375239968

Change-Id: Ibee4cd8cf12f7909c31314e4dfe3586d13a7886b
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5958528
Commit-Queue: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Pin-chih Lin <johnylin@google.com>