CHROMIUM: iwl7000: mac80211: wake the queues in case of failure in resume
In case we fail to resume, we'll WARN with
"Hardware became unavailable during restart." and we'll wait until user
space does something. It'll typically bring the interface down and up to
recover. This won't work though because the queues are still stopped on
IEEE80211_QUEUE_STOP_REASON_SUSPEND reason.
Make sure we clear that reason so that we give a chance to the recovery
to succeed.

Change-Id: I76a15fdb92de450e5329940125f3c58916be3942
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: d93c0b68c9504f2af4a7c8a9233c743e92dffd26