CHROMIUM: arm64: dts: mt8186: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading
the power, the display power is calculated in watts due to
unsynchronized reading and writing. Use virtual battery 0xf
to communicate with EC can solve this.

BUG=b:361198692
TEST=check battery charging and discharge function
UPSTREAM-TASK=b:355547083

Change-Id: Ie0c2fb72e32338f17add74a37274d8659009e381
Signed-off-by: Frank Wu <frank_wu@compal.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6024667
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Shawn Ku <shawnku@chromium.org>
Commit-Queue: Ivan Chen <yulunchen@google.com>
Reviewed-by: Ivan Chen <yulunchen@google.com>