CHROMIUM: arm64: dts: mt8186: Add ponyta360.dtsi to adapt
Considering that there may be different SKU configurations in the future,
ponyta360 will be built separately to distinguish.

Since it is inherited from steelix.dtsi, touchscreen@5d is not used in
this project and occupies the same group of control tchpins, which may
cause touchscreen@41 to fail to register. disabled the 5d node to fix it.

BUG=b:375297946
UPSTREAM-TASK=b:389549480
TEST=emerge-staryu sys-kernel/chromeos-kernel-5_15

Change-Id: I988e3d596ce153174deb655efb1941de6106d1e7
Signed-off-by: Haikun Zhou <zhouhaikun5@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6174279
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Shawn Ku <shawnku@chromium.org>
Reviewed-by: Zhengqiao Xia <xiazhengqiao@huaqin.corp-partner.google.com>
Commit-Queue: Shawn Ku <shawnku@chromium.org>